back/requirements.txt
django
djangorestframework
psycopg2-binary
requests
django-cors-headers
gunicorn
python-dotenv